Implemented loyalty points calculation upon purchase

+++ online-bookstore/online-bookstore-domain/src/main/java/es/pedrazamiguez/assessment/onlinebookstore/domain/model/PurchaseContext.java
@@ -18,7 +18,7 @@ public class PurchaseContext {
   private boolean paymentProcessed;
   private String shippingAddress;
   private boolean shippingPrepared;
-  private int loyaltyPointsEarned;
+  private Long loyaltyPointsEarned;
 
   public boolean isSuccessful() {
     return this.status == PurchaseStatus.SUCCESS || this.status == PurchaseStatus.PENDING;
